Add logger, Add basic config, Add sync log loading
Add logger, A centalized logger like this lets us setup information we
require to be logged because we use structured logging. Add config, the
config type gives us the ability to more fine tune the config in code as
well as in the config file. The Load config done this way decides the
log level and a few other things. Add start function in index, this is
done so we can syncronously read the config in while it being async. Its
a dumb byproduct of async await.