Add Slash Command Registerer
Slash commands have to be registered with discord. They have to be
registered so they can have the nice pretty fancy UI. They can be
registered either globally(on all guild the bot is on) or per guild
basis. Globally takes an hour at most to take effect because it has to
stay in cache. Per guild takes effect immediately on that specific
server but not at all on the others.